refactor(pricing): split resolver into parse/fallback/resolve submodules
resolver.rs had grown to 807 lines (past the U-16 soft limit of 800) after
the CN vendor work plus rustfmt's method-chain wrapping. Split it into a
resolver/ directory: parse.rs (parse_litellm_data), fallback.rs
(fallback_pricing + openai/xai helpers), resolve.rs (resolve_pricing_known),
with mod.rs re-exporting the three functions so db.rs's import path is
unchanged.

Pure refactor: no behavior change. Tests pass unchanged (431 lib / 43
integration / 2 sdk), clippy and fmt clean. Largest file is now parse.rs
at 350 lines.
